+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+# Unreleased
+* If an error occurs when when setting up video recording, disable video recording instead of crashing
+
 # 1.0.0 (2020-07-12)
 * First release